Create function.go linter
